[modi] 목업 데이터 정렬
